refactor: split protocols
