+++ broker/src/main/java/io/camunda/zeebe/broker/transport/backupapi/BackupApiRequestHandler.java
@@ -13,8 +13,8 @@ import io.camunda.zeebe.broker.transport.ErrorResponseWriter;
 import io.camunda.zeebe.logstreams.log.LogStreamRecordWriter;
 import io.camunda.zeebe.protocol.impl.record.RecordMetadata;
 import io.camunda.zeebe.protocol.impl.record.value.management.CheckpointRecord;
-import io.camunda.zeebe.protocol.record.AdminRequestType;
-import io.camunda.zeebe.protocol.record.BackupRequestType;
+import io.camunda.zeebe.protocol.management.AdminRequestType;
+import io.camunda.zeebe.protocol.management.BackupRequestType;
 import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.management.CheckpointIntent;
