+++ broker/src/main/java/io/camunda/zeebe/broker/transport/partitionapi/InterPartitionCommandSenderImpl.java
@@ -13,8 +13,8 @@ import io.camunda.zeebe.backup.api.CheckpointListener;
 import io.camunda.zeebe.backup.processing.state.CheckpointState;
 import io.camunda.zeebe.broker.Loggers;
 import io.camunda.zeebe.broker.partitioning.topology.TopologyPartitionListenerImpl;
-import io.camunda.zeebe.clustering.management.InterPartitionMessageEncoder;
-import io.camunda.zeebe.clustering.management.MessageHeaderEncoder;
+import io.camunda.zeebe.broker.protocol.InterPartitionMessageEncoder;
+import io.camunda.zeebe.broker.protocol.MessageHeaderEncoder;
 import io.camunda.zeebe.engine.transport.InterPartitionCommandSender;
 import io.camunda.zeebe.protocol.record.ValueType;
 import io.camunda.zeebe.protocol.record.intent.Intent;
