+++ broker/src/main/resources/broker-protocol.xml
@@ -1,26 +1,10 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
 <sbe:messageSchema xmlns:sbe="http://fixprotocol.io/2016/sbe"
   xmlns:xi="http://www.w3.org/2001/XInclude"
-  package="io.camunda.zeebe.clustering.management" id="5" version="2"
-  semanticVersion="0.1.0" description="Zeebe Cluster Management Protocol" byteOrder="littleEndian">
+  package="io.camunda.zeebe.broker.protocol" id="5" version="2"
+  semanticVersion="0.1.0" description="Zeebe Broker Communication Protocol" byteOrder="littleEndian" >
 
   <xi:include href="../../../../protocol/src/main/resources/common-types.xml"/>
-
-  <types>
-
-    <!-- binary data -->
-    <composite name="blob">
-      <type name="length" primitiveType="uint32" maxValue="2147483647"/>
-      <type name="varData" primitiveType="uint8" length="0"/>
-    </composite>
-
-    <enum name="errorResponseCode" encodingType="uint8" semanticType="String">
-      <validValue name="PARTITION_NOT_FOUND">0</validValue>
-      <validValue name="INVALID_PARAMETERS">1</validValue>
-      <validValue name="READ_ERROR">2</validValue>
-    </enum>
-
-  </types>
 
   <sbe:message id="1" name="ExporterPositions">
     <group id="0" name="positions">
@@ -36,7 +20,7 @@
     <field name="recordKey" id="3" type="uint64" presence="optional"/>
     <field name="checkpointId" id="4" type="int64"/>
 
-    <data name="command" id="32" type="blob"/>
+    <data name="command" id="32" type="varDataEncoding"/>
   </sbe:message>
 
 </sbe:messageSchema>
