+++ gateway/src/main/java/io/camunda/zeebe/gateway/admin/BrokerAdminRequest.java
@@ -11,9 +11,9 @@ import io.camunda.zeebe.gateway.impl.broker.request.BrokerRequest;
 import io.camunda.zeebe.gateway.impl.broker.response.BrokerResponse;
 import io.camunda.zeebe.protocol.impl.encoding.AdminRequest;
 import io.camunda.zeebe.protocol.impl.encoding.AdminResponse;
-import io.camunda.zeebe.protocol.record.AdminRequestEncoder;
-import io.camunda.zeebe.protocol.record.AdminRequestType;
-import io.camunda.zeebe.protocol.record.AdminResponseEncoder;
+import io.camunda.zeebe.protocol.management.AdminRequestEncoder;
+import io.camunda.zeebe.protocol.management.AdminRequestType;
+import io.camunda.zeebe.protocol.management.AdminResponseEncoder;
 import io.camunda.zeebe.transport.RequestType;
 import io.camunda.zeebe.util.buffer.BufferWriter;
 import java.util.Optional;
