+++ protocol/src/main/resources/cluster-management-protocol.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<sbe:messageSchema xmlns:sbe="http://fixprotocol.io/2016/sbe"
+  xmlns:xi="http://www.w3.org/2001/XInclude" package="io.camunda.zeebe.protocol.management"
+  id="1" version="0" semanticVersion="${project.version}"
+  description="Zeebe Cluster Management Protocol" byteOrder="littleEndian">
+
+  <xi:include href="common-types.xml"/>
+
+  <types>
+    <enum name="AdminRequestType" encodingType="uint8">
+      <validValue name="STEP_DOWN_IF_NOT_PRIMARY">0</validValue>
+      <validValue name="PAUSE_EXPORTING">1</validValue>
+    </enum>
+
+    <enum name="BackupRequestType" encodingType="uint8">
+      <validValue name="TAKE_BACKUP">0</validValue>
+    </enum>
+  </types>
+
+  <sbe:message name="AdminRequest" id="1">
+    <field name="partitionId" id="1" type="uint16"/>
+    <field name="type" id="2" type="AdminRequestType"/>
+    <field name="brokerId" id="3" type="uint16" presence="optional"/>
+  </sbe:message>
+
+  <sbe:message name="AdminResponse" id="2">
+  </sbe:message>
+
+  <sbe:message name="BackupRequest" id="3">
+    <field name="partitionId" id="1" type="uint16"/>
+    <field name="type" id="2" type="BackupRequestType"/>
+    <field name="backupId" id = "3" type="int64"/>
+  </sbe:message>
+
+  <sbe:message name="BackupResponse" id="4">
+  </sbe:message>
+</sbe:messageSchema>
