+++ protocol/src/main/resources/protocol.xml
@@ -80,14 +80,6 @@
       <type name="patchVersion" primitiveType="int32"/>
     </composite>
 
-    <enum name="AdminRequestType" encodingType="uint8">
-      <validValue name="STEP_DOWN_IF_NOT_PRIMARY">0</validValue>
-      <validValue name="PAUSE_EXPORTING">1</validValue>
-    </enum>
-
-    <enum name="BackupRequestType" encodingType="uint8">
-      <validValue name="TAKE_BACKUP">0</validValue>
-    </enum>
 
   </types>
 
@@ -129,23 +121,6 @@
     <data name="bpmnProcessId" id="1" type="varDataEncoding"/>
   </sbe:message>
 
-  <sbe:message name="AdminRequest" id="40">
-    <field name="partitionId" id="1" type="uint16"/>
-    <field name="type" id="2" type="AdminRequestType"/>
-    <field name="brokerId" id="3" type="uint16" presence="optional" sinceVersion="4"/>
-  </sbe:message>
-
-  <sbe:message name="AdminResponse" id="41">
-  </sbe:message>
-
-  <sbe:message name="BackupRequest" id="42">
-    <field name="partitionId" id="1" type="uint16"/>
-    <field name="type" id="2" type="BackupRequestType"/>
-    <field name="backupId" id = "3" type="int64"/>
-  </sbe:message>
-
-  <sbe:message name="BackupResponse" id="43">
-  </sbe:message>
 
   <!-- L2 Common Messages 200 - 399 -->
